src/requirements.txt
opencv-python==4.6.0.66
RPi.GPIO==0.7.1